Implemented save method in the helper
We now store the buffer as a property in our
bitmaps. When you save a bitmap, it writes
this buffer back to the supplied filename.
NOTE: We currently have arrays for colors and
pixels, but changes to those arrays will not
alter the underlying buffer. We need to change
our implementation to deal with this.