Moved buffer access into getter methods
The constructor method was not ideal, because by creating
color and pixel arrays as properties, we were duplicating
that data in memory. Not only that, but it would increase
the complexity of the save operation to the point where
our helper would have to be a bit smarter. The goal is
to keep all the logic in the bitmap constructor file, and
use the buffer directly inside.